+++ src/react/components/editabletable/XUIEditableTable.js
@@ -259,7 +259,7 @@ XUIEditableTable.propTypes = {
    *
    * Recommended English value:
    * <br />
-   * *Press space bar to start a drag. When dragging you can use the arrow keys to move the item around and escape to cancel. Ensure your screen reader is in focus mode or forms mode.*
+   * *Press Space bar or Enter to start a drag. When dragging you can use the arrow keys to move the item around and escape to cancel. Ensure your screen reader is in focus mode or to use your pass through key.*
    */
   dndInstructions(props, propName, componentName) {
     return conditionallyRequiredValidator(